pokemon/pokemon_backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
django-filter==23.5
coreapi==2.3.3 